back/pyproject.toml
[tool.poetry]
name = "datagotchi"
version = "0.1.0"
description = "Datagotchi backend API"
authors = ["Datagotchi Team <team@datagotchi.ai>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic = "^2.7.1"
python-dotenv = "^1.0.1"
notte-sdk = "*"
pydantic-settings = "^2.9.1"
pandas = "^2.2.2"
pyarrow = "^15.0.2"
Pillow = "^11.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^24.3.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 